+++ deenurp/test/test_subcommand_filter_outliers.py
@@ -4,7 +4,7 @@ from Bio import SeqIO
 
 from deenurp import wrap
 from deenurp.subcommands import filter_outliers
-from deenurp.util import which, MissingDependencyError
+from deenurp.util import which
 
 from deenurp.test import util
 
@@ -23,9 +23,17 @@ class FilterOutliersFunctions(unittest.TestCase):
         self.assertEqual(len(seqnames), distmat.shape[0])
 
     @unittest.skipUnless(which(wrap.USEARCH), "{} not found.".format(wrap.USEARCH))
-    def test_distmat_usearch(self):
+    def test_distmat_pairwise_usearch(self):
         infile = util.data_path('e_faecalis.head.fasta')
-        taxa, distmat = filter_outliers.distmat_usearch(infile, 'foo')
+        taxa, distmat = filter_outliers.distmat_pairwise(
+            infile, 'foo', 'usearch', wrap.USEARCH)
+        self.assertEqual(distmat.shape[0], len(taxa))
+
+    @unittest.skipUnless(which(wrap.VSEARCH), "{} not found.".format(wrap.VSEARCH))
+    def test_distmat_pairwise_vsearch(self):
+        infile = util.data_path('e_faecalis.head.fasta')
+        taxa, distmat = filter_outliers.distmat_pairwise(
+            infile, 'foo', 'vsearch', wrap.VSEARCH)
         self.assertEqual(distmat.shape[0], len(taxa))